+++ src/cloudant/document.py
@@ -16,7 +16,6 @@
 API module/class for interacting with a document in a database.
 """
 import json
-import posixpath
 import requests
 from requests.exceptions import HTTPError
 
@@ -87,19 +86,19 @@ class Document(dict):
 
         # handle design document url
         if self._document_id.startswith('_design/'):
-            return posixpath.join(
+            return '/'.join((
                 self._database_host,
                 url_quote_plus(self._database_name),
                 '_design',
                 url_quote(self._document_id[8:], safe='')
-            )
+            ))
 
         # handle document url
-        return posixpath.join(
+        return '/'.join((
             self._database_host,
             url_quote_plus(self._database_name),
             url_quote(self._document_id, safe='')
-        )
+        ))
 
     def exists(self):
         """
@@ -389,7 +388,7 @@ class Document(dict):
         """
         # need latest rev
         self.fetch()
-        attachment_url = posixpath.join(self.document_url, attachment)
+        attachment_url = '/'.join((self.document_url, attachment))
         if headers is None:
             headers = {'If-Match': self['_rev']}
         else:
@@ -432,7 +431,7 @@ class Document(dict):
         """
         # need latest rev
         self.fetch()
-        attachment_url = posixpath.join(self.document_url, attachment)
+        attachment_url = '/'.join((self.document_url, attachment))
         if headers is None:
             headers = {'If-Match': self['_rev']}
         else:
@@ -473,7 +472,7 @@ class Document(dict):
         """
         # need latest rev
         self.fetch()
-        attachment_url = posixpath.join(self.document_url, attachment)
+        attachment_url = '/'.join((self.document_url, attachment))
         if headers is None:
             headers = {
                 'If-Match': self['_rev'],